+++ .gitignore
@@ -242,6 +242,4 @@ $RECYCLE.BIN/
 
 # End of https://www.toptal.com/developers/gitignore/api/windows,macos,linux,visualstudiocode,node
 /db*
-
-.env
-/db
+/config*
