+++ controllers/workController.js
@@ -23,3 +23,5 @@ const feed = async (req, res) => {
 };
 
 module.exports = { worksList, feed };
+
+module.exports = { uploadImages };
