+++ middlewares/util.js
@@ -1,0 +1,19 @@
+const util = {
+  success: (status, message, data) => {
+    return {
+      status: status,
+      success: true,
+      message: message,
+      data:data
+    }
+  },
+  fail: (status, message) => {
+    return {
+      status: status,
+      success: false,
+      message: message
+    }
+  }
+}
+
+module.exports = {util};
